Provide PHP enumerations that can perform the appropriate checks themselves.
I considered implementing the logic within the `CoreFunctions` but opted to keep all of the verification logic self-contained within the enumerations themselves.

This allows for the reduction of the extra type parameter.
The enumeration knows its own type and can perform the appropriate check.

This is designed to ensure that all appropriate checks are performed in order to guarantee that a boolean is returned.
The caller need only perform the `is()` or the `in()` check using the appropriate configuration type.

Only the most simplistic types are chosen.
If additional php `is_*()` types are needed, then enumerations may be added.

The special `NULL` supporting types are done to reduce the needed checks should something care to check if the type is either NULL or a given type.
This doesn't increase the code complexity much at all and the use should reduce code complexity.
The `NULL` checks need to be performed anyway.

A specific type `ANY` can be used to test if a configuration type exists as is NULL.
The `in()` method allows for array key existence checks and then a type check on some child property.
This should allow for shallow array value type checks.

The non-`$config` based `isset()` and `is_string()` checks are reverted to previous behavior.

This should allow for the configuration checks to be performed anywhere without warnings or problems.